pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rulecrawl"
version = "1.0.0"
description = "A visual, rule-based crawler system with 5-tab node orchestration."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["crawler", "spider", "visual", "fastapi", "mongodb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "motor==3.6.0",
    "pydantic==2.9.0",
    "httpx==0.27.0",
    "lxml==5.3.0",
    "cssselect==1.2.0",
    "parsel==1.9.0",
    "jsonpath-ng==1.6.1",
    "python-dotenv==1.0.1"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/rulecrawl"
"Bug Tracker" = "https://github.com/yourusername/rulecrawl/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
